api/API_Hibrida/Respaldos/API_Python/hospital-farmacia-python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
aiomysql==0.2.0
motor==3.6.0
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.12
